Part 4.21

+++ part4/bloglist/controllers/blogs.js
@@ -37,13 +37,21 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  const result = await Blog.findByIdAndRemove(request.params.id)
-  
-  if (result) {
-    response.status(204).end()
-  } else {
-    response.status(404).end()
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
   }
+
+  const userid = decodedToken.id
+  const blog = await Blog.findById(request.params.id)
+
+  if (blog.user.toString() === userid.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    return response.status(204).end()
+  }
+  return response.status(404).end()
 })
 
 blogsRouter.put('/:id', async (request, response) => {
